requirements.txt
apify-client>=1.7.0
requests>=2.31.0
python-dotenv>=1.0.0
anthropic>=0.40.0
